+++ injectool/__init__.py
@@ -2,6 +2,7 @@
 
 __version__ = '1.1.2'
 
-from .core import DependencyError, Container, SingletonResolver, FunctionResolver
-from .core import make_default, add_resolver, add_singleton, add_function_resolver, resolve
+from .core import DependencyError, Resolver, Container
+from .resolvers import SingletonResolver, FunctionResolver
+from .resolvers import add_resolver, add_singleton, add_function_resolver, add_type
 from .injection import inject, dependency